requirements.txt
pandas
sqlalchemy
psycopg2-binary
requests
apache-airflow
python-dotenv
apache-airflow-providers-postgres